requirements.txt

fastapi>=0.110.0
uvicorn>=0.27.0
gunicorn==21.2.0

groq>=0.9.0
google-genai>=0.3.0

supabase>=2.4.0

python-dotenv>=1.0.1
pydantic>=2.6.0

pillow>=10.2.0
python-multipart>=0.0.9